refactor(permissions): Modify name and refactor code for simplicity
