+++ app/core/permissions.py
@@ -2,17 +2,17 @@ from flask import request, g
 import jwt
 from functools import wraps
 from app.model.models import User
-from app.model.sessions import get_session
 from app.core.exceptions.app_errors import UnauthorizedError, ForbiddenError
 from typing import Callable, Any
 
-from flask import current_app
+from app.core.auth_utils import decode_token
 
 # Décorateur Authentification (token JWT)
 def auth_required(func: Callable) -> Callable:
     '''
     Vérifie la présence et la validité du token
     dans l'en-tête `Authorization: Bearer <token>`.
+    Utilise g.session pour les requêtes SQLAlchemy.
 
     Retourne une fonction décorée appliquant la vérification JWT.
     Lève une erreur si token expiré, invalide ou manquant.
@@ -23,28 +23,18 @@ def auth_required(func: Callable) -> Callable:
         if not auth_header or not auth_header.startswith("Bearer "):
             raise UnauthorizedError("Token manquant")
 
-        # !!! Extraire la fonction de vérification et de decodage token !!!
-        JWT_KEY = current_app.config.get("JWT_KEY")
-        JWT_ALGO = current_app.config.get("ALGORITHM", "HS256")
         token = auth_header.split(" ")[1]
-        try:
-            current_user = jwt.decode(token, JWT_KEY, JWT_ALGO)
-        except jwt.ExpiredSignatureError:
-            raise UnauthorizedError("Token expiré")
-        except jwt.InvalidTokenError:
-            raise UnauthorizedError("Token invalide")
-
-        user_id = current_user.get("id")
+        payload = decode_token(token)
+        user_id = payload.get("id")
         if not user_id:
             raise UnauthorizedError("Payload Token invalide")
 
-        session = get_session()
+        session = g.session
         try:
             user = session.get(User, user_id)
             if not user:
                 raise UnauthorizedError("Utilisateur introuvable")
 
-            # Ajouter g.current_user: absent -> echec sur "commandes" !
             g.current_user = user
 
             return func(current_user=user, *args, **kwargs)
