+++ app/routes/order_routes.py
@@ -1,5 +1,5 @@
 from flask import request, Blueprint, jsonify, g, Response
-from app.core.auth import access_granted
+from app.core.permissions import access_granted
 from app.services.order_utils import (
     get_all_orders, get_order_by_id, create_new_order,
     change_status_order, get_orderitems_all
